+++ ui/src/shared/icons/LogoIcon.svelte
