+++ ui/src/shared/icons/PocketIcon.svelte
